Feat: Add token refresh to MangaDex API
This commit adds automatic token refreshing to the MangaDex API implementation.
API calls that return a 401 Unauthorized response will now attempt to refresh the access token using the stored refresh token. If successful, the API call is retried.

It also includes a minor update to the `MangaLibraryGrid` component to adjust the Coil image loading scale based on whether a poster image is available.

Signed-off-by: AtelierMizumi <thuanc177@gmail.com>